Stop using Object.defineProperty in resultMapper.
I respectfully disagree with the comment about hiding the ID field, since
the property is still discoverable in principle. If we really wanted to
keep hiding it, we could do that only in development.

Simple property assignment is not only faster than calling
Object.defineProperty, but also makes subsequent usage of the object
faster. Kind of a shame that defineProperty is still slow in 2018, but
that's the world we live in.